requirements.txt
streamlit
pypdf2
langchain
python-dotenv
faiss-cpu
huggingface_hub
sentence_transformers
InstructorEmbedding
#-e .